Fix bug after press '=' button
The bug occurs when the user tries to insert a number after the function equalButton. Instead of the display shows the new number, it would shows a string with the result from the function equalButton plus the number inserted by the user.
The expected behavior is to erase the result from the function and only shows the new number added by the user.

The equalButton function only update the display with the result from operate() and pass an empty string to firstNumber variable.
The operatorButtons function updates the firstNumber variable with the display content, in the case of the user tries to realize an operation with the result in the display.